requirements.txt
# Project
flask
python-dotenv

## OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
